move plotting.surface._utils related tests to plotting.surface.tests.test_utils
